+++ saleor/tests/fixtures.py
@@ -8,6 +8,7 @@ from typing import List, Optional
 from unittest.mock import MagicMock, Mock
 
 import pytest
+import pytz
 from django.conf import settings
 from django.contrib.auth.models import Group, Permission
 from django.contrib.sites.models import Site
@@ -16,6 +17,7 @@ from django.core.files.uploadedfile import SimpleUploadedFile
 from django.db import connection
 from django.forms import ModelForm
 from django.test.utils import CaptureQueriesContext as BaseCaptureQueriesContext
+from django.utils import timezone
 from django_countries import countries
 from PIL import Image
 from prices import Money, TaxedMoney
@@ -26,7 +28,10 @@ from ..app.types import AppType
 from ..checkout import utils
 from ..checkout.models import Checkout
 from ..checkout.utils import add_variant_to_checkout
+from ..core import JobStatus
 from ..core.payments import PaymentInterface
+from ..csv.events import ExportEvents
+from ..csv.models import ExportEvent, ExportFile
 from ..discount import DiscountInfo, DiscountValueType, VoucherType
 from ..discount.models import (
     Sale,
@@ -36,6 +41,7 @@ from ..discount.models import (
     VoucherTranslation,
 )
 from ..giftcard.models import GiftCard
+from ..invoice.models import Invoice
 from ..menu.models import Menu, MenuItem, MenuItemTranslation
 from ..order import OrderStatus
 from ..order.actions import cancel_fulfillment, fulfill_order_line
@@ -45,6 +51,7 @@ from ..order.utils import recalculate_order
 from ..page.models import Page, PageTranslation
 from ..payment import ChargeStatus, TransactionKind
 from ..payment.models import Payment
+from ..plugins.invoicing.plugin import InvoicingPlugin
 from ..plugins.models import PluginConfiguration
 from ..plugins.vatlayer.plugin import VatlayerPlugin
 from ..product import AttributeInputType
@@ -327,6 +334,21 @@ def checkout_with_voucher_percentage_and_shipping(
     checkout.shipping_method = shipping_method
     checkout.shipping_address = address
     checkout.save()
+    return checkout
+
+
+@pytest.fixture
+def checkout_with_payments(checkout):
+    Payment.objects.bulk_create(
+        [
+            Payment(
+                gateway="mirumee.payments.dummy", is_active=True, checkout=checkout
+            ),
+            Payment(
+                gateway="mirumee.payments.dummy", is_active=False, checkout=checkout
+            ),
+        ]
+    )
     return checkout
 
 
@@ -1373,6 +1395,12 @@ def order_events(order):
 @pytest.fixture
 def fulfilled_order(order_with_lines):
     order = order_with_lines
+    invoice = order.invoices.create(
+        url="http://www.example.com/invoice.pdf",
+        number="01/12/2020/TEST",
+        created=datetime.datetime.now(tz=pytz.utc),
+        status=JobStatus.SUCCESS,
+    )
     fulfillment = order.fulfillments.create(tracking_number="123")
     line_1 = order.lines.first()
     stock_1 = line_1.allocations.get().stock
@@ -2251,3 +2279,78 @@ def app_installation():
         app_name="External App", manifest_url="http://localhost:3000/manifest",
     )
     return app_installation
+
+
+@pytest.fixture
+def user_export_file(staff_user):
+    job = ExportFile.objects.create(user=staff_user)
+    return job
+
+
+@pytest.fixture
+def app_export_file(app):
+    job = ExportFile.objects.create(app=app)
+    return job
+
+
+@pytest.fixture
+def export_file_list(staff_user):
+    export_file_list = list(
+        ExportFile.objects.bulk_create(
+            [
+                ExportFile(user=staff_user),
+                ExportFile(user=staff_user,),
+                ExportFile(user=staff_user, status=JobStatus.SUCCESS,),
+                ExportFile(user=staff_user, status=JobStatus.SUCCESS),
+                ExportFile(user=staff_user, status=JobStatus.FAILED,),
+            ]
+        )
+    )
+
+    updated_date = datetime.datetime(
+        2019, 4, 18, tzinfo=timezone.get_current_timezone()
+    )
+    created_date = datetime.datetime(
+        2019, 4, 10, tzinfo=timezone.get_current_timezone()
+    )
+    new_created_and_updated_dates = [
+        (created_date, updated_date),
+        (created_date, updated_date + datetime.timedelta(hours=2)),
+        (
+            created_date + datetime.timedelta(hours=2),
+            updated_date - datetime.timedelta(days=2),
+        ),
+        (created_date - datetime.timedelta(days=2), updated_date),
+        (
+            created_date - datetime.timedelta(days=5),
+            updated_date - datetime.timedelta(days=5),
+        ),
+    ]
+    for counter, export_file in enumerate(export_file_list):
+        created, updated = new_created_and_updated_dates[counter]
+        export_file.created_at = created
+        export_file.updated_at = updated
+
+    ExportFile.objects.bulk_update(export_file_list, ["created_at", "updated_at"])
+
+    return export_file_list
+
+
+@pytest.fixture
+def user_export_event(user_export_file):
+    return ExportEvent.objects.create(
+        type=ExportEvents.EXPORT_FAILED,
+        export_file=user_export_file,
+        user=user_export_file.user,
+        parameters={"message": "Example error message"},
+    )
+
+
+@pytest.fixture
+def app_export_event(app_export_file):
+    return ExportEvent.objects.create(
+        type=ExportEvents.EXPORT_FAILED,
+        export_file=app_export_file,
+        app=app_export_file.app,
+        parameters={"message": "Example error message"},
+    )